% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Btree.R
\docType{methods}
\name{Btree}
\alias{Btree}
\alias{Btree,ANY-method}
\title{"Birch Algorithm"}
\usage{
Btree(x, ...)

\S4method{Btree}{ANY}(x)
}
\arguments{
\item{x}{the input data}
}
\description{
This package implements a the Birch algorithm using HDF5 data
}
\author{
"Aparo Antonino" "Da Col Fabio" "Purrone Alberto"
}
